backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
azure-search-documents>=11.6.0
azure-storage-blob>=12.20.0
azure-ai-vision-imageanalysis>=1.0.0
openai>=1.35.0
azure-identity>=1.17.0
aiohttp>=3.9.0
httpx>=0.27.0
Pillow>=10.3.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
